chore :truck: Files referring to get-availabilities renamed
