+++ README.md
@@ -6,9 +6,9 @@
 - [x] Deve ser possível fazer o cadastro como consultant
 - [x] Deve ser possível fazer login como user
 - [x] Deve ser possível fazer login como consultant
-- [ ] Deve ser possível listar todos as horários de um consultant
-- [ ] Deve ser possível listar todas as reuniões de um usuário
-- [ ] Deve ser possivel listar todas as reuniões de um consultant
+- [x] Deve ser possível listar todos as horários de um consultant
+- [x] Deve ser possível listar todas as reuniões de um usuário
+- [x] Deve ser possivel listar todas as reuniões de um consultant
 - [ ] Deve ser possivel listar todas as reuniões de um determinado usuário e consultant
 - [x] Consultants devem poder cadastrar horários
 - [x] Consultants devem poder poder deletar horários
