+++ src/use-cases/factories/make-list-consultant-availabilities-use-case.ts
@@ -1,9 +1,11 @@
 import { PrismaAvailabilityRepository } from "@/repositories/prisma/prisma-availability-repository";
-import { GetAvailabilitiesUseCase } from "../availability/get-availabilities";
+import { ListConsultantAvailabilitiesUseCase } from "../availability/list-consultant-availabilities";
 
 export function makeGetAvailabilitiesUseCase() {
   const availabilityRepository = new PrismaAvailabilityRepository();
-  const useCase = new GetAvailabilitiesUseCase(availabilityRepository);
+  const useCase = new ListConsultantAvailabilitiesUseCase(
+    availabilityRepository
+  );
 
   return useCase;
 }
